backend/mysite/requirements.txt
Django==3.0.5
django-cors-headers==3.2.1
django-filter==2.2.0
djangorestframework==3.11.0
SQLAlchemy==1.1.13
sqlparse==0.3.1
